Rename Period to Interval in JS
