+++ app/templates/DonationForm.html.twig
@@ -183,7 +183,7 @@
                         stateKey: 'donationFormContent.addressType'
                     },
                     {
-                        viewHandler: WMDE.View.createPaymentPeriodAndAmountDisplayHandler(
+                        viewHandler: WMDE.View.createPaymentIntervalAndAmountDisplayHandler(
                                 $( '.interval-text' ),
                                 $( '.amount-formatted'),
                                 {
